src/write_to_mysql/requirements.txt
pyspark
mysql-connector-python
cassandra-driver
requests
kafka-python